[fastapi] fix regression in type conversion
crrev.com/c/5832361 changed FastApiCall to be typed using the signature
instead of always using Tagged. It also added the assumption that
fallbacks must throw if the arguments do not match how V8 interprets
them for the fast call. This added an ambiguity with string handling
because "is a string" is enough for a fallback to but the fast call
will still bail out due to it. This ambiguity was resolved by disabling
fast calls to functions that use the OneByteString type, which resulted
in regressions for embedders.

This CL handles fallback return values so that OneByteString (and
everything else) can be used properly.

BYPASS_LARGE_CHANGE_WARNING Most new code is in tests.

Bug: 41497260
Change-Id: Ibbbf98843a68c0b812b1c4a834625e5532d5bf36
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5870687
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96714}