[base] Support uint32_t in FormattedString
The `FormattedString` so far supported the `int` and `size_t` types.
Adding more unsigned types turned out to be tricky, because depending on
the platform the types `uint32_t`, `uint64_t`, and `size_t` can be the
same type.

So instead, this CL allows all integers (of 32 or 64-bit, but this can
be extended if needed) and of either signedness. This is implemented in
one generic `FormattedStringPart`.

R=mliedtke@chromium.org

Bug: 367752285
Change-Id: I10db8344be4d0c9e11ea959b1fb17f64caa10b3f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5952968
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96744}