[cfi] Use fine-grained RWX scopes for debugging
We need all writes to RWX memory to go through our API so that we can
add validation. However, the scope objects currently unlock the JIT and
thereby allow any writes of the code that is using them.
Add a mode for debug builds in which we don't unlock the JIT and instead
only unlock it for the writes in our API itself.

Bug: 42203297
Change-Id: Ie8362217720846c2fe509a30eaa3412f0bbbbf6a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5953725
Commit-Queue: Stephen Röttger <sroettger@google.com>
Reviewed-by: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96747}