[riscv][wasm] Fix Load/Store trapping in LiftoffAssembler
5929007 fixed Load/Store trapping in CodeGenerator. This one fixes the
same problem in LiftoffAssembler. In few words: exact position of a
Load/Store instruction to be trapped cannot be calculated outside of
MacroAssembler, so we pass trapping routine to MASM as a functor and so
let MASM call it right in place.

Change-Id: I605c378855b423c04db6a1d3b4f7378edb9e1ccb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5953144
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Reviewed-by: Yahan Lu <yahan@iscas.ac.cn>
Commit-Queue: Yahan Lu <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#96817}