[fuzzer] Initialize compilation state for reference module
The compilation state needs to be initialized in case any "regular"
compilation is triggered for the reference module.
While the reference module is somewhat special, it should still be in
a state that is as close to "production" code as possible.

Not having the compilation state initialized can result in DCHECKs or
crashes in case the reference execution triggers a deopt.
(Regular compilations expect the compilation_progress_ to have an entry
per function defined in the module which is not the case if the
compilation state is not initialized accordingly.

Bug: 373924119, 42204618
Change-Id: I668001ad75f228cdb46c10d2850b9ec502b2c454
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5952529
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96766}