[collections] Support big WeakMaps
The hashes stored in JSReceivers are maximally 20
bits. If a hash table is bigger than that, all
items get clumped in the beginning.

This CL distributes the probes more evenly in big
hash tables. It's still not optimal (there will be
empty space because of quadratic probing) but it's
better than the previous solution.

Bug: 42207427,375033082

Change-Id: I9a49a52e8369255c1536151c970b8b8524f4e8b4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5946964
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96776}