[heap] Use visitor-dispatch in GC instead of instance type-dispatch
This CL moves visitor towards visitor_id-dispatch for traversing
objects instead of the instance type-based dispatch. This is in order
to unify V8 on a single dispatch mechanism.

For the most part this CL only needs to change the super class of
visitors from ObjectVisitorWithCages to HeapVisitor. In addition
this CL had to add the following changes:

Add ShouldUseUncheckedCast() in HeapVisitor for large new space
objects in Scavenger. The DCHECK inside the Cast() would fail
otherwise because the right map word isn't in-place.

The visitor-dispatch is now used in places where the object size
was already computed in advance. So this CL also allows to pass that
size to the visitor as an argument and uses it internally in the
VisitXXX()-methods. Visitors without a pre-computed size still need
to compute the size internally. In order to make this less
error-prone, this CL introduces MaybeObjectSize which doesn't have a
value set for such visitors.

Bug: 372749729
Change-Id: Icafa7997ca508d94bbbce30fc85f755b95cd2b99
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5940376
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96784}