Reland: [collections] Support big WeakMaps
The hashes stored in JSReceivers are maximally 20
bits. If a hash table is bigger than that, all
items get clumped in the beginning.

This CL distributes the probes more evenly in big
hash tables. It's still not optimal (there will be
empty space because of quadratic probing) but it's
better than the previous solution.

Side fix: "are we running out of capacity" computation
in RehashObjectHashTableAndGCIfNeeded was too pessimistic
(it added slack twice, first there and once in
ComputeCapacity).

Bug: 42207427,375033082
Change-Id: Ic53052421459165b4cdecc3052fb8bf480c9373d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5961516
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96798}