[sandbox] Refactor code-compatibility check in JSDispatchTable
The sandbox requires basic control-flow integrity for JavaScript
function calls: (1) the target code must use JS linkage, and (2) the
caller and callee must agree on the signature. With the JSDispatchTable,
these requirements become simple invariants that can explicitly be
checked when creating/updating a table entry: the code must use the JS
code entrypoint tag and its parameter count must be compatible with that
of the table entry.

This CL refactors these checks a bit by moving the logic into a helper
function (IsCompatibleCode). This then avoids some code duplication and
makes the code a bit more readable. The checks are now also a bit more
strict: previously, we allowed any varargs code to be installed on
non-varargs dispatch entries. Now we only allow that for certain
builtins for which we know that they can correctly deal with this.

Drive-By: Correctly mark some builtin functions as varargs, i.e. as not
requiring argument adaption.

Bug: 337547182, 40931165
Change-Id: I41e05b338b249b5e6fb4c9d3ebd6d0c1769565d4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5961339
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96806}