[wasm] Remove NativeModule argument on wrapper compilation
Wrapper compilation is module independent. The `NativeModule` was only
used to get the `CompilationEnv`, and there was already a fallback to
use the `NoModuleAllFeatures` environment.
We also remove that fallback, and do not pass any environment to wrapper
compilation. The only used fields are enabled features, so we just use
`WasmEnabledFeatures::All()` inside wrapper compilation.

R=jkummerow@chromium.org

Bug: 375343414
Change-Id: Iaf546fa12924c02a7164bca5f12018f99bc842b6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5962127
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96831}