[turbofan] Refine keyed feedback with known named keys
This is a straightforward port of this Maglev CL:
https://crrev.com/c/5953688.

This allows us to optimize o[k] if the site is megamorphic (helper
function called with multiple k) but every call itself has a known k.

generally useful because ThinStrings are not always safe to be read in
the background, but is the Actual is an InternalizedString, then it
will always be safe to read in the background (which could thus open
more optimization opportunities).

Drive-by: bypass ThinString in HeapConstants in Turbofan. This is
Change-Id: Ib4d430d2482bf0166a8cacfcd134c44575877e2d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5965076
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96857}