[wasm][jspi] Simplify stack limit change
We don't need to set the other limit if one of the CAS fails, it will be
set by the interrupting thread, and the other CAS ensures that we don't
overwrite it.
Also remove some dead code that handled setting the stack limit from a
runtime call. This is no longer used since the function for changing the
stack limit for stack switching was replaced with a fast C call, which
does not need to switch to the central stack.

R=jkummerow@chromium.org,clemensb@chromium.org
CC=irezvov@chromium.org

Bug: 355400671
Change-Id: I508d3817d3e332bf3b253235bfc4a0ec6e076dee
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5966735
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96888}