[shadow stack] Build shadow stack in deoptimization
Deoptimizer builds stack frames including return addresses, but shadow
stack does not have the added return addresses. Therefore, it causes
security exception due to the mismatch when CET shadow stack is enabled.

This CL builds a valid shadow stack during the deoptimization.
`AdaptShadowStackForDeopt`, which is called at the end
DeoptimizationEntry, jumps to the previous instruction of each return
address which is `call AdaptShadowStackForDeopt`. Therefore, it
recursively pushes the return address onto the shadow stack.

This implementation is based on the following doc.
https://docs.google.com/presentation/d/1maphEn0mDldQkbbYnvv13dxl9BbHNwq0Q6iTXrkd_Iw

Also, since we push the addresses onto the stack, and jump to them
while building shadow stack, we must validate the addresses to
prevent attackers from using this code and jumping to arbitrary code.
This CL reuses `IsValidReturnAddress` in deoptimizer-cfi-builtins.
It makes a crash if it's not one of the addresses in the allow list.

WASM deopt is not supported yet in this CL.

Bug: chromium:42201233
Change-Id: I93acd2279838f225de562509ae6660c0759cda67
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5855923
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Choongwoo Han <choongwoo.han@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#96965}