[cppgc] Print debug information if marking verifier fails
In debug builds or with cppgc_enable_verify_heap = true, the marking
verifier may fail if the total size of the objects that were found to
be marked in all pages is not equal to the expected total size of
marked objects. In the unlikely case that this happens, this CL
outputs a list of the objects (marked and unmarked) that were found
on the pages for which there was a mismatch between the expected size
of marked objects and the actual size. This output may be helpful for
debugging purposes, especially when cppgc_enable_object_names = true
is also used.

Change-Id: I60979cdacca28805805c1c152cfc3ea784994826
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5898695
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96869}