[wasm][jspi] Drop climit for native builds
The stack guard's climit is only relevant for simulator builds, which
have two separate stacks for JS and C++.
Drop it for native builds and always use the JS limit instead. This
avoids having to update both when updating the stack limit, which is
performance sensitive in the context of wasm JSPI/stack-switching.

R=clemensb@chromium.org,ishell@chromium.org

Bug: 355400671
Change-Id: I9bd369bc70bb50251bedc919a951490a81d0255f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5973174
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96893}