[maglev] fix phi untagging bug with single-block loop phis
This CL fixes a bug with phi untagging a single-block loops. Here is
an example of when this bug could happen. Consider initially the
following Phis (taken from the attached repro):
```
│  21: φᵀⁱ a0 (n9, n23) (compressed) → (x), 7 uses
│  22: φᵀ r0 (n2, n21) (compressed) → (x), 2 uses
│  23: φᵀⁱ r1 (n20, n31) (compressed) → (x), 7 uses
```

Note how 23 is an input to 21. After Phi untagging, we have

```
│  21: φᴵⁱ a0 (n10, n48) → (x), 8 uses
│  22: φᵀ r0 (n2, n49) (compressed) → (x), 2 uses
│  23: φᴵⁱ r1 (n16, n28) → (x), 7 uses
...
│  48: CheckedSmiUntag [n23:(x)] → (x), 1 uses
```

As you can see, 48 expects n23 to be Tagged, but it's untagged.

48 was actually inserted here
https://source.chromium.org/chromium/chromium/src/+/main:v8/src/maglev/maglev-phi-representation-selector.cc;l=530-534;drc=c6ad4dc9eab118aec32079cd0fe9996385205760
when untagging 21. Then, we decided to untag 23. What should have then
happened is that we should have eventually visited 48 and called
UpdateNodeInputs and then UpdateUntaggingOfPhi, which would have
converted the CheckedSmiUntag into a CheckedSmiSizedInt32. However,
when we insert new nodes in the current block, we delay their
insertion until we've reached the end of the current block (because we
don't want to be modifying the node list while we iterate it), and we
never actually call UpdateNodeInputs on them. In particular for such
loop phi backedge untagging nodes, this is usually not an issue,
because they are created when visiting the loop header, but inserted
at the end of the backedge block, and usually, header block !=
backedge block. However, here, the loop contains a single block, hence
the issue.

The fix (this CL) is pretty simple: inserting the nodes right away
(which happens to actually be safe, so the whole "delaying the
insertion to avoid modifying the node list while iterating it" thing
was actually useless from the start), so that UpdateNodeInputs gets
called on them since they will be picked up by the regular node
iterator.

Bug: v8:7700
Change-Id: I82336b7d3aa29727b5b5b3f347b8e876fab6d6a3
Fixed: chromium:375330032
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5970433
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96883}