[regexp] Implement global mode for experimental
.. and fix the kTreatMatchAtEndAsFailure quirk implementation.

For the former, the experimental engine already supports global mode and
it was simply a matter of threading everything through.

For the latter, we incorrectly looked at the start of the first match
instead of the last match.

Bug: 370804671, 372285209
Change-Id: I6592cd52283dad551e38fef611ae7f1ded9c63ac
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5974325
Commit-Queue: Patrick Thier <pthier@chromium.org>
Auto-Submit: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96915}