Remove "robust" string traversal
We expect strings to be "robust" (for traversal) by default now and it's
unclear that the rudimentary sanity checking performed by LooksValid
would really help if an invalid string was being processes as it only
looks at the owning MemoryChunk. As such, this CL removes both the
RobustnessFlag and the LooksValid() helper method. This then also
prepares for further changes to the String::ToCString function.

Bug: 376071292
Change-Id: I193e94eb260e88163304a9671e6ff88ee73e9c79
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5979832
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96919}