[regexp] Port `split` to the global irregexp API
Ideally this would be able to use the batched / global mode, but due
to non-standard AdvanceStringIndex use this isn't trivial.

Likewise, the LastMatchInfo should really only be set for the last
result. But this is hard because we have to ignore any empty match at
the end of the string.

We do a plain port over to the new API instead. the RegExpExec quirks
are now unused and can be removed.

Drive-by: Fix an issue with setting LastMatchInfo in
RegExpExecInternal_Batched in which it was possible to use a stale
offsets vector pointer.

Bug: 372285209
Change-Id: Ie09774167d3b7c14013ed93214ddd678ba437067
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5979834
Auto-Submit: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96931}