Move expected parameter count into scratch register on arm64
The debug-only code to validate that optimized functions see a valid
dispatch handle compares the parameter count of the dispatch table entry
with the expected one. Previously, the expected parameter count was an
immediate in the cmp instruction, however, that fails as soon as the
parameter count becomes too large to fit into 12 bits. With this CL, the
parameter count is now first moved into a scratch register.

Bug: 376786828
Change-Id: Ib684477ac2ebb5e1cb821c3a373b6c765a2d68b2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5987670
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96986}