[wasm] Refactor wrapper tracking in WasmDispatchTables
It turns out that having to keep calls to WasmDispatchTable::Set
and WasmDispatchTableData::Add in sync is too brittle, so this
CL refactors them to be a single call that handles everything.

This fixes a bug where we got it wrong: when creating a
WasmDispatchTable for an existing WasmTableObject in
InitializeImportedIndirectFunctionTable, we must account for any
wrappers taken over from existing dispatch tables.

Fixed: 375343420
Change-Id: I609b348f88f69a4235b7388cbf24d6f687d4e119
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5971454
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96966}