[turboshaft][arm] Remove arm orderfile workaround
We believe the issue on the arm32 orderfile bot to be caused by (or a
combination with) the context snapshot trial which is now properly
disabled on 32 bit arm. This CL reverts the initial workaround in the
Turboshaft builtin pipeline (--turboshaft-csa) to verify this.

Bug: 325104859
Change-Id: I497ac9f1778d5fbfcfee90519180c88919870dc7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5993570
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96970}