[sandbox] Add regression tests for String::WriteToFlat
This is a follow-up for for https://crrev.com/c/5332375 which adds some
regression tests to ensure that String::WriteToFlat does not write
out-of-bounds. This CL also adds some more documentation and DCHECKs
into String::WriteToFlat to document some of the assumptions around the
handling of cons strings.

Bug: 369748453
Change-Id: Ifc07c4595316ffd17edc34fd7bfe68e58bc4b01e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5979529
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#96983}