[regexp] Fix reentrancy wrt the result offsets vector
.. by ensuring that only a single regexp execution uses the static
result offsets vector at a time. We do this by adding one level of
indirection, i.e.: an isolate field stores the ptr to the static
vector, and is set to nullptr when the vector is in use. When the
static vector is unavailable, we use dynamic allocation.

Fixed: 370804671
Change-Id: I34d2280f30e4bac9750d8bea3bbaa49c52b07009
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5993587
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97007}