[sandbox] Don't rely on in-heap flags in DeleteMemoryChunk
In DeleteMemoryChunk we need to determine the concrete
MemoryChunkMetadata subtype at runtime. Previously, we used
"metadata->Chunk()->IsLargePage()" for that purpose. That, however, ends
up relying on in-heap flags which should not be trusted. With this CL we
now introduce a IsLargePage accessor in the MutablePageMetadata class
and use that instead to determine the correct type.

Bug: 332855890
Change-Id: I24ed7841f48eeacf6759581c9085b26453619b16
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999473
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97015}