[objects] Remove GetHeap() from JSObjectShortPrint()
JSObjectShortPrint() had special detection for some broken
references. This CL removes this because it doesn't provide much
value anymore.

The GetHeap() call was using GetHeapFromWritableObject() under
the hood. This was triggering DCHECKs for shared objects and
because of that all JS objects in the shared heap needed special
handling in JSObjectShortPrint().

Bug: 376564609
Change-Id: If0a5cb56e70cac19bfd71c9f4ab027075731d19d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999919
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97034}