[sandbox] Ensure correct state of PageMetadata objects during evacuation
During GC evacuation, we currently rely on the in-sandbox MemoryChunk
flags to determine what to do with a given page. If these flags are
corrupted, we might end up in a confused state, causing for example a
UaF of a PageMetatdata object. This CL adds a SBXCHECK to protect
against that. Longer term, we could investigate having trusted page
flags (see https://crbug.com/377724745) to avoid similar issues.

Bug: 352481087
Change-Id: Iffb80f8fa36407146c325312e977f642049bea38
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6000162
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97038}