[scopes] Reenable scope info reuse without hard check
Instead when we find a duplicate we do some simple verification that
the scope infos match. If this works we'll know that the failing
CHECK_EQ comes from an equivalent scope info that we didn't manage to
merge. If it fails we'll know that our UniqueIds aren't in fact
Unique; possibly because we fall of the end of eval or some other
synthetic scope info.

Change-Id: I03a6e9f38ce9ce87afc9b4541991f7ebdf626571
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999548
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97057}