[float16array] Use better f64->f16 codegen where supported
This CL also changes the TruncateFloat64ToFloat16 machine operator
to output a uint16, because there is currently no float16
representation, and outputting a uint16 is the only existing use case
(for converting on storing into Float16Array). It is simpler than
adding a float16 representation to the compilers and a separate
BitcastFloat16ToWord32 operator.

Bug: 42203953
Change-Id: If21c1b6590165bfa5412abae5faf699dac196353
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6001554
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Ilya Rezvov <irezvov@chromium.org>
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97107}