Prepare Android GPU FYI for O deprecation
This prepares https://crrev.com/c/5994945 which drops the device
name from the builder and updates the device type.

The two CLs will land as much in sync as possible, but there'll
be a short window where the builder is broken until both are in.
The builder doesn't block lkgr.

Bug: 377389499
Change-Id: If2cb2a3e3fd98be42bf2cc58ec55d5d868768a7c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999578
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>