[string] Use size_t for output length of String::ToCString
String::ToCString returns an off-heap char[] allocated on the C++ heap,
so the buffer's size should use size_t instead of uint32_t. In general
we want to use size_t for the sizes of objects outside of the V8 heap
and uint32_t for sizes of objects inside the V8 heap (for example to
ensure that corruption of size values can not cause an out-of-sandbox
memory access later on).

Bug: 376071292
Change-Id: I7692b7b4b0e93d63d93b13c8104e1ea9900e3ab1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6000364
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97101}