[sandbox] Try to reserve the first four GB of the address space
This is useful to mitigate Smi<->HeapObject confusion issues where a
(32-bit) Smi is treated as a pointer and dereferenced. For now, we
simply try a few times to reserve this part of the address space
starting at different offsets (as most operating systems forbid
allocation the null page and a few pages afterwards) and give up if that
doesn't work. The V8::IsSandboxConfiguredSecurely API can be used to
determine whether the reservation was successful.

Bug: 40925855
Change-Id: Ia727509497f0c8bba063c4209079f743b71edc92
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999723
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97067}