[scopeinfo] Drop check that we're not updating old sfis
Either we're off the main thread in which case we're not touching the
old sfis, or we're on the main thread and may already have installed
the new sfi in the old sfi table. In the latter case the DCHECK was
broken.

Change-Id: I0a0841eaa2f9c4b54689a36015df731607bc26b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6010567
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97084}