Reland "[turbofan] Fix bug with StringWrapper named access"
This is a reland of commit 5585128538b6609b9eeab20c07615a5c24ec926c.

The issue original CL was that for non-Load AccessMode, Receiver and
lookup_start_object should be the same, but this wasn't the case
because we were introducing the TypeGuard only on lookup_start_object
and not on receiver.

Original CL description:
> [turbofan] Fix bug with StringWrapper named access
>
> Fixed: chromium:377213712
> Change-Id: Icf9d45a4c4fc20d664da6c24ff9b58b23b2da816
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999501
> Reviewed-by: Toon Verwaest <verwaest@chromium.org>
> Commit-Queue: Toon Verwaest <verwaest@chromium.org>
> Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#97068}

Bug: chromium:377213712
Change-Id: Idc37a9ba34793cada8d8635032d318adf9bce2fc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6006536
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97091}