[api] Use unsigned length values in V8::String Value types
This CL switches lengths of V8::String::Value and V8::String::ValueView
to uint32_t and the length of V8::String::Utf8Value to size_t. This is
both a follow-up for https://crrev.com/c/5332375 (which converted the
internal String::length_ to an uint32_t) and preparation for
https://crrev.com/c/5975682 which will introduce a new API for Utf8
encoding a String that uses size_t.

Bug: 369748453, 373485796
Change-Id: I156a18aa3b0e421dc1b0dc3741c8f2877144071e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6012365
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97111}