[maglev] Ensure smi/number check before storing to script context slot
If the property’s script context slot is expected to contain a
kSmi, we must ensure that only Smi values are stored in this slot.
Any non-Smi value will trigger a deoptimization.

If it is expected to be kMutableHeapNumber, then it accepts
any number (either Smi or HeapNumber).
By emitting a StoreDoubleField operation, the input value is
automatically converted to float64. The conversion might
be done by unsafe nodes, since we emit CheckNumber.

Drive-by: refactor TrySpecializeStoreScriptContextSlot.

Fixed: 377850462
Change-Id: I4a304a8ebfbb268b7664b1f776deed9ecc48d6b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6012358
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97122}