Add unit tests for the v8 ETW provider and implement support for capturing state when a URL filter is supplied while enabling the trace.
To implement unit tests for controlling the v8 ETW provider, I
introduced an indirect way of calling into the Isolate via a new
EtwIsolateOperations type.

I also needed to access the EtwIsolateLoadScriptData type from the
tests, so I relocated it out of the anonymous namespace it was in to
its own file and marked the type V8_EXPORT_PRIVATE.  The only
changes to the methods of this type are to introduce
EtwIsolateOperations and to update EnableLogWithFilterData so that
it clears the cache of previously loaded scripts to ensure they
are reemitted when handling a request to capture state.

Bug: 340621497
Change-Id: I090d6b97daa8bd1b3b0f9d0cfab26d8a611e09a2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5936871
Reviewed-by: Seth Brenith <seth.brenith@microsoft.com>
Commit-Queue: Bo Cupp <pcupp@microsoft.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97171}