Implement a fuzz test for SharedMutexThreads
This makes the test SharedMutexThreadsFuzz shared between its
original test instance (which logic is unchanged) and a fuzz test
that enables fuzzing this via the fuzz-test engine. The fuzz test
will be packaged as part of our fuzz-test executable and the target
will get some fuzzing time on Clusterfuzz.

See also: https://github.com/google/fuzztest

This transition also serves as a blueprint for similar transitions
of viable tests to fuzz test.

Providing such tests as fuzz tests will make it easier in the
future to explore non-deterministic test behavior and reduce such
behavior in normal testing. This can help to make tests less flaky
or faster. In the case of this test, it isn't flaky (didn't fail
yet), but it's not fast for a unit test due to the busy waiting
in the test.

Bug: 378073354
Change-Id: I8c21a410ea47a086d6d02de1f1a7b72a9ed34686
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5962126
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97113}