[string][sandbox] Add internal::String::WriteUtf8 API
This CL introduces a new API for Utf8-encoding a V8 string into an
(external) buffer. Subsequently, this new internal API will be used for
the (new) public V8::String::WriteUtf8 API, as well as potentially in
other places in V8 that currently manually implement Utf8 encoding.

The main motivation for the new API is to ensure sandbox compatibility:
the old APIs (implicitly) assumed that a string's length could never
change. As such, they for example allowed passing -1 as output length,
meaning "until the end of the string". This can now (almost) never be
safe, as the string's length may change in between computing the size of
the output buffer and encoding the string into it. With the new API, the
caller now always needs to specify the capacity of the output buffer,
and we never write more than that many bytes. Furthermore, if null
termination is requested, the result will now always be null terminated,
even if that means encoding one fewer character from the string.

Bug: 376071292
Change-Id: I2b0dbda85dc5f56e1ea630cbd33b664c3f4b69fc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6012361
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97126}