[wasm] Canonical SIMD shuffle matching
The 'ArchShuffles', for the x64 backend, have already been moved into
the wasm directory for mid-end optimisations to use. I would like to
do the same for AArch64, and since there is a lot of duplication, it
made sense to add a generic API to match specific shuffle patterns.

Nothing is hooked up to use this yet, other than tests.

Change-Id: I046519f1e8c8e15e43c8b3c3994d7b2ac0806d14
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5999499
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97145}