[maglev] Reload old context slot value in StoreScriptContextSlotWithWriteBarrier
Previously, we passed the old value as an input to this node,
assuming that it would likely be cached due to the common pattern
of V8 emitting an LdaContextSlot just before a StaContextSlot.

However, if the slot contains a mutable heap number and context
specialization is not enabled, this approach fails. In this case,
we end up allocating a new heap number, meaning the cached old
value may not represent the actual heap number in the context slot.
We would then crash when trying to mutate the old value.

Importantly, when context specialization is enabled, we wouldn’t
emit StoreScriptContextSlotWithWriteBarrier at all.

This brings the code closer to TF, aligning with the behavior in
StoreScriptContextSlot reduction in js-typed-lowering.cc.

Fixed: 377966014
Change-Id: I70e8e6fc9841a34c284d352973bca71d7fea5889
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6010766
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97148}