[sandbox] Change behavior of --sandbox-testing mode
The --sandbox-testing mode is used for demonstrating and validation V8
sandbox issues. Previously, this mode required demonstrating the ability
to perform an arbitrary write by writing to a specific page outside of
the sandbox. With this change, this requirement is relaxed and instead
any memory corruption outside of the sandbox will be reported as a
"sandbox violation".

With that, --sandbox-testing and --sandbox-fuzzing are now mostly
identical, with only a few minor differences:
* --sandbox-fuzzing requires the memory corruption API to be enabled
  (mostly as a sanity check) while --sandbox-testing can be enabled on
  any build.
* --sandbox-testing will exit with status zero (indicating success) when
  detecting harmless crashes while --sandbox-fuzzing will exit with a
  non-zero status instead. This way, fuzzers can still detect premature
  process termination while tests will be treated as passing if they for
  example fail a CHECK or otherwise cause a safe crash.

Note that both --sandbox-testing and --sandbox-fuzzing will also report
out-of-sandbox _reads_ as sandbox violations. This is not technically
true as the sandbox only attempts to prevent _writes_. However it is
both difficult and undesirable to filter out reads, for example because
in some cases, the underlying issue may also allow performing a write.

This change then also updates a number of regression tests that
previously relied on the (now removed) Sandbox.targetPage address.

Bug: 42202821
Change-Id: Idf34aa73a487052f4cafb9740c0d9ceaeea60201
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6012455
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97151}