[fastapi][test] Remove CHECK in testing code
The fast API test function enforce_range_compare_xxx had a CHECK that
required that two parameters have the same values. The fuzzer naturally
found values where the two parameters were not the same. This CL cleans
up the implementation and tests of enforce_range_compare_xxx so that an
exception gets thrown when a value is out of range.

The CL also removes the `in_range`, which does not seem to be necessary
anymore. The CL also adjust the tests such that the `in_range` parameter
is not used anymore, and it cleaned up some parameters that are not
representable as numbers.

R=marja@chromium.org

Bug: 373924020
Change-Id: If79ad66c3c5103e6db220331a58a02cd2da845da
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5946966
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97152}