[string] Fix String::Flatten wrt special TF 'flat' cons strings
We ran into a problem with a special form of 'flat' cons strings
as created by TF, in which cons->first is empty (instead of the
canonical form in which cons->second is empty). Such strings were
NOT flattened in-place by String::Flatten. Indeed, these were the
only string shape for which in-place flattening did not occur.

RegExpExecInternal in its current form relies on in-place
flattening due to how the ToDirectStringAssembler works.

This CL changes String::Flatten to canonicalize such special TF
'flat' cons strings.

 // The old contract:
 s2 = String::Flatten(s1); CHECK(s2->IsFlat());
 // The new contract:
 s2 = String::Flatten(s1); CHECK(s1->IsFlat() && s2->IsFlat());

Drive-by: Micro-optimizations. Move SlowFlatten to the -inl header,
and reduce handle dereferences.

Bug: 372285209
Fixed: 377829476, 378102651, 378014586
Change-Id: I5225b5aed81b94cc7063518a1c607ff3ffc86a94
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6012196
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97181}